+++ frontend/jupiter_front/src/pages/MyPage.js
@@ -1,7 +1,13 @@
+import axios from "axios";
+
 const MyPage = () => {
+
     return (
         <>
         나의 페이지 영역
+        {sessionStorage.getItem('email')}
+        {/* //email:"aaa@abc.com", userid:"user1", username: "일반사용자" */}
+
         </>
     )
 }
